moving and renaming files to have a cleaner file tree
