+++ src/lerobot/policies/fastwam/wan/README.md
@@ -1,6 +1,10 @@
-# Wan2.2 Upstream Subset
+# FastWAM `wan` package
 
-This directory contains the trimmed subset of the official Wan2.2 source tree used by FastWAM.
+This package holds FastWAM's model implementation. It mixes a small **vendored
+subset of the official Wan2.2 source tree** with FastWAM's own code, kept flat in
+a single directory.
+
+## Vendored from Wan2.2
 
 - Upstream repository: https://github.com/Wan-Video/Wan2.2
 - Upstream commit: `42bf4cfaa384bc21833865abc2f9e6c0e67233dc`
@@ -8,18 +12,22 @@ This directory contains the trimmed subset of the official Wan2.2 source tree us
 
 Copied files:
 
-- `wan/modules/attention.py`
-- `wan/modules/model.py`
-- `wan/modules/__init__.py`
-- `wan/utils/fm_solvers.py`
-- `wan/utils/__init__.py`
+- `attention.py` (was `wan/modules/attention.py`)
+- `model.py` (was `wan/modules/model.py`)
+- `get_sampling_sigmas` in `video_dit.py` (was `wan/utils/fm_solvers.py`), inlined
+  next to its only caller.
 
-This subset now only backs FastWAM's **custom MoT video DiT**. The Wan2.2 VAE,
-UMT5 text encoder, and tokenizer are no longer vendored — they come from
+This subset only backs FastWAM's **custom MoT video DiT**. The Wan2.2 VAE,
+UMT5 text encoder, and tokenizer are no longer vendored - they come from
 `diffusers.AutoencoderKLWan`, `transformers.UMT5EncoderModel`, and
-`transformers.AutoTokenizer` (see `../wan_adapters.py` and `../wan_components.py`).
+`transformers.AutoTokenizer` (see `components.py` and `adapters.py`).
 
-Current FastWAM adapters that directly reuse this vendored subset:
+## FastWAM's own code
 
-- `../wan_video_dit.py` builds on `wan.modules.model` (`sinusoidal_embedding_1d`, `rope_params`, `rope_apply`, …) and `wan.modules.attention.flash_attention`.
-- `../modular_fastwam.py` reuses `wan.utils.fm_solvers.get_sampling_sigmas` for Wan-compatible inference timesteps.
+- `video_dit.py` builds on `model` (`sinusoidal_embedding_1d`, `rope_params`,
+  `rope_apply`, …) and `attention.flash_attention`. Its
+  `WanContinuousFlowMatchScheduler` uses `get_sampling_sigmas` for Wan-compatible
+  inference timesteps.
+- `components.py` / `adapters.py` load the VAE, text encoder, tokenizer, and the
+  custom DiT weights.
+- `modular.py` defines the FastWAM model (`ActionDiT`, `MoT`, `FastWAM`, …).
