--- src/lerobot/policies/fastwam/wan/modules/__init__.py
@@ -1,8 +1,0 @@
-# Copyright 2024-2025 The Alibaba Wan Team Authors. All rights reserved.
-from .attention import flash_attention
-from .model import WanModel
-
-__all__ = [
-    "WanModel",
-    "flash_attention",
-]
