--- src/lerobot/policies/fastwam/wan/utils/__init__.py
@@ -1,6 +1,0 @@
-# Copyright 2024-2025 The Alibaba Wan Team Authors. All rights reserved.
-from .fm_solvers import get_sampling_sigmas
-
-__all__ = [
-    "get_sampling_sigmas",
-]
